3d_matching_eval/requirements.txt
matplotlib>=3.5.1
numba>=0.55.1
numpy>=1.21.5
opencv-python-headless>=4.5.5.64
pandas>=1.3.5
python-dotenv>=0.20.0
scipy>=1.7.3
typer>=0.4.1
PyYaml>=6.0
pycpd>=2.0.0
